pyproject.toml
[tool.poetry]
name = "learn-django"
version = "0.1.0"
description = ""
authors = ["MahirSalahin <mahirsalahin01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
psycopg2 = "^2.9.9"
factory-boy = "^3.3.0"
faker = "^26.0.0"
graphene-django = "^3.2.2"
django-filter = "^24.2"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^4.4.6"
drf-nested-routers = "^0.94.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
